refactor:#37 재사용할 컴포넌트 파일 위치 이동
